TEST - Add Show

+++ routes/web.php
@@ -16,10 +16,10 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', 'PageController@index')->name('welcome');
 
 //7 routes
-Route::get('comics', 'Admin\ComicController@index')->name('comics');
-Route::get('comics/create', 'Admin\ComicController@create')->name('admin.comics.create');
-Route::post('comics', 'Admin\ComicController@store')->name('admin.comics.store');
-Route::get('comics/{comic}', 'Admin\ComicController@show');
-Route::get('comics/{comic}/edit', 'Admin\ComicController@edit');
-Route::put('comics/{comic}', 'Admin\ComicController@update');
-Route::delete('comics/{comic}', 'Admin\ComicController@destroy');
+Route::get('admin/comics', 'Admin\ComicController@index')->name('comics');
+Route::get('admin/comics/create', 'Admin\ComicController@create')->name('admin.comics.create');
+Route::post('admin/comics', 'Admin\ComicController@store')->name('admin.comics.store');
+Route::get('admin/comics/{comic}', 'Admin\ComicController@show')->name('admin.comics.show');
+Route::get('admin/comics/{comic}/edit', 'Admin\ComicController@edit')->name('admin.comics.edit');
+Route::put('admin/comics/{comic}', 'Admin\ComicController@update')->name('admin.comics.update');
+Route::delete('admin/comics/{comic}', 'Admin\ComicController@destroy')->name('admin.comics.destroy');
